Themed
Update to Bootstrap 3 and moved all of the templates (the "theme") to an
external theme package (sculpin/bootstrap-3-blog-theme) to showcase the new
theme functionality.

There might still be some hiccups, but I think that most of the functionality
that was originally available in the blog skeleton are still intact.